requirements.txt
streamlit==1.32.0
pandas==2.2.1
google-generativeai==0.3.2
python-dotenv==1.0.1
XlsxWriter==3.1.9 